requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
duckdb
apache-airflow-providers-mysql
sqlalchemy
apache-airflow-providers-amazon
